update CAN packet priorities
If two devices attempt to send at the same time, the second one will
interrupt the first and listeners will focus on the second transmission.
However, the last packet of the first transmission has a higher priority
and will corrupt the second transmission. Therefore we exchange the
priorities of normal data packets and the last packet of each
transmission.